Move baz tests out of example dir
